Update android tests.
