+++ android/app/src/androidTestGps/java/org/pathcheck/covidsafepaths/gps/storage/RealmSecureStorageTest.kt
@@ -1,10 +1,12 @@
-package org.pathcheck.covidsafepaths.storage
+package org.pathcheck.covidsafepaths.gps.storage
 
 import com.facebook.react.bridge.ReadableArray
 import com.facebook.react.bridge.WritableMap
 import com.facebook.react.bridge.WritableNativeArray
 import com.facebook.react.bridge.WritableNativeMap
 import com.marianhello.bgloc.data.BackgroundLocation
+import covidsafepaths.gps.storage.Location
+import covidsafepaths.gps.storage.RealmSecureStorage
 import io.realm.Realm
 import io.realm.kotlin.where
 import org.junit.After
