+++ android/app/src/androidTestGps/org/pathcheck/covidsafepaths/DetoxTest.java
