+++ android/app/src/bte/java/covidsafepaths/bte/MainApplication.java
@@ -1,47 +1,40 @@
-package covidsafepaths;
+package covidsafepaths.bte;
 
 import android.app.Application;
 import android.content.Context;
 import com.facebook.react.PackageList;
 import com.facebook.react.ReactApplication;
-import com.marianhello.bgloc.BackgroundGeolocationFacade;
 import com.facebook.react.ReactNativeHost;
 import com.facebook.react.ReactPackage;
 import com.facebook.soloader.SoLoader;
-import org.pathcheck.covidsafepaths.BuildConfig;
-import io.realm.Realm;
 import java.lang.reflect.InvocationTargetException;
 import java.util.List;
-import covidsafepaths.bridge.RealmPackage;
-import covidsafepaths.storage.SecureStorage;
+import org.pathcheck.covidsafepaths.BuildConfig;
 
 public class MainApplication extends Application implements ReactApplication {
 
   private static Context context;
 
   private final ReactNativeHost mReactNativeHost =
-          new ReactNativeHost(this) {
-            @Override
-            public boolean getUseDeveloperSupport() {
-              return BuildConfig.DEBUG;
-            }
+      new ReactNativeHost(this) {
+        @Override
+        public boolean getUseDeveloperSupport() {
+          return BuildConfig.DEBUG;
+        }
 
-            @Override
-            protected List<ReactPackage> getPackages() {
-              @SuppressWarnings("UnnecessaryLocalVariable")
-              List<ReactPackage> packages = new PackageList(this).getPackages();
-              // Packages that cannot be autolinked yet can be added manually here, for example:
-              // packages.add(new MyReactNativePackage());
-              packages.add(new RealmPackage());
+        @Override
+        protected List<ReactPackage> getPackages() {
+          @SuppressWarnings("UnnecessaryLocalVariable")
+          List<ReactPackage> packages = new PackageList(this).getPackages();
 
-              return packages;
-            }
+          return packages;
+        }
 
-            @Override
-            protected String getJSMainModuleName() {
-              return "index";
-            }
-          };
+        @Override
+        protected String getJSMainModuleName() {
+          return "index";
+        }
+      };
 
   @Override
   public ReactNativeHost getReactNativeHost() {
@@ -54,10 +47,6 @@ public class MainApplication extends Application implements ReactApplication {
     MainApplication.context = getApplicationContext();
     SoLoader.init(this, /* native exopackage */ false);
     initializeFlipper(this); // Remove this line if you don't want Flipper enabled
-    Realm.init(this);
-    initializeGeolocationTransformer();
-//    // Ignore assignment. Creating to begin heavy encryption work
-    SecureStorage wrapper = SecureStorage.INSTANCE;
   }
 
   /**
@@ -84,17 +73,6 @@ public class MainApplication extends Application implements ReactApplication {
         e.printStackTrace();
       }
     }
-  }
-
-  private void initializeGeolocationTransformer() {
-    BackgroundGeolocationFacade.setLocationTransform((context, location) -> {
-
-      // Save Location in encrypted realm db
-      SecureStorage.INSTANCE.saveDeviceLocation(location);
-
-      // Always return null. We never want to store data in the libraries SQLite DB
-      return null;
-    });
   }
 
   public static Context getContext() {
