+++ android/app/src/gps/java/covidsafepaths/gps/MainActivity.java
@@ -1,4 +1,4 @@
-package covidsafepaths;
+package covidsafepaths.gps;
 
 import android.os.Bundle;
