+++ android/app/src/gps/java/covidsafepaths/gps/bridge/RealmPackage.kt
@@ -1,4 +1,4 @@
-package covidsafepaths.bridge
+package covidsafepaths.gps.bridge
 
 import android.view.View
 import com.facebook.react.ReactPackage
