+++ android/app/src/gps/java/covidsafepaths/gps/bridge/SecureStorageManager.kt
@@ -1,13 +1,13 @@
-package covidsafepaths.bridge
+package covidsafepaths.gps.bridge
 
 import com.facebook.react.bridge.Promise
 import com.facebook.react.bridge.ReactApplicationContext
 import com.facebook.react.bridge.ReactContextBaseJavaModule
 import com.facebook.react.bridge.ReactMethod
 import com.facebook.react.bridge.ReadableArray
-import covidsafepaths.storage.Location.Companion.SOURCE_GOOGLE
-import covidsafepaths.storage.Location.Companion.SOURCE_MIGRATION
-import covidsafepaths.storage.SecureStorage
+import covidsafepaths.gps.storage.Location.Companion.SOURCE_GOOGLE
+import covidsafepaths.gps.storage.Location.Companion.SOURCE_MIGRATION
+import covidsafepaths.gps.storage.SecureStorage
 
 class SecureStorageManager(reactContext: ReactApplicationContext) : ReactContextBaseJavaModule(reactContext) {
